This update includes an initial rewrite of OO queries for both speed and API improvements.  The api may change a little over the next few commits as I find bugs or missing features.  The biggest change is in the join syntax.  There are now five join methods on queries and iterators:  join, leftJoin, rightJoin, fullJoin, innerJoin.  The type argument has therefore been removed.  Join is shorthand for leftJoin.
The new syntax for join is: join(joinFromObjectAlias, joinToObjectAlias, relationshipAlias, alias).

joinFromObjectAlias - I am the alias of the object being joined from.  
joinToObjectAlias - I am the alias of the object being joined to.
relationshipAlias - I am the alias of the relationship to use when joining these two objects.
alias - I the alias of the object in the query.

The record.xsl was updated.  This means that generated records need to be deleted and recreated.

Also, query pooling has temporarily been removed as it is suspect for the current memory leak problems and the new object structure seems to be almost as fast based on simple timer tests.

Various attribute names have been changed for clarity too.

Lastly, the samples apps (except for scratch) have been updated to work with this version.

The documentation is still out of date.