This update continues the trend of making the identification of relationships more explicit.  There were a few issues with iterators related to this.  In particular, when you were saving the contents of iterator which was a child of another object.
For example,  User.save() causes all the addresses in the user's addressIterator to save (if they're dirty).  There were some problems with iterators in the case that a user had multiple relationships with addresses.  Reactor didn't know which relationship to use when saving.  

Additionally, the xsl for records was update to explicitly identify the relationship.

Lastly, I've been working to clean up the API of the metadata objects. Due to this a few poorly implemented and unneeded methods were removed.