Fixed a bug with the last commit (reactor.query.query != reactor.query.object)
Also implemented delete methods on the gateways per ticket #5.  This added the following methods to all gateways:

deleteAll() - this deletes all rows in the table.
deleteByFields(...) - this accepts name/value pairs to match and delete.
deleteByQuery(query) - this accepts a query object to delete by.  Note: the query can not have any joins and the select list is ignored.