Fixes #14.  Ticket 14 identified a problem with decimal (and numeric) sql data types being truncated by the cfqueryparam tag.  The suggestion resolution was to update all of the adapters (ObjectDao.cfc files) to inspect the database for scale and to use that to set the scale attribute of the cfqueryparam tags throughout the framework.
Though technically feasible, it turned out to be an impractical solution.  The scale attribute applies to any cf_sql_numeric cfsqltype and must be greater than 0 (despite the fact that that's a valid (and default) option for ms SQL server (at least)).  This meant that I would need to check each cfqueryparam tag to see if they needed the scale attribute before applying it.  This is easy in the XSL files, but not so easy in the abstractGateway.

An alternative was simply to change cf_sql_decimal cfsqltypes to cf_sql_float.  It turns out that the scale attrbiute simply tells CF where to truncate a decimal value before it's passed to the DBMS.  This means that if I specify the data type as cf_sql_float that CF won't try to truncate the value and the DBMS will take care of it.  

Furthermore, the focus of cfqueryparams in Reactor serves specifically two purposes: 

1) security - to help prevent SQL injection attacks
2) prepared statements - to improve performance

I don't see any way that specifiy the scale attribute would improve security and the queries should still become perpared statements in the DBMS.