Fixes #96 and Fixes #87.
The ObjectDao.cfc files all now add an additional item of data introspected from the database:  readonly.

If the DAO, for whatever reason, determines that a row should be readonly then this property is set to true.

Any field which is read only will be reflected in the metadata field information.  Also, the dao will not insert or update this field.  Lastly, records will not get a setter for the field.

This resolves issues with the rowGuid created by SQL server for replication and for computed columns.  

All of the ObjectDaos (except for SQL Server) don't take anything into consideration.  Right now, SQL Server's dao only looks at rowguid and computed columns.

An email will be sent to the mailing list asking for contributions where needed in the other Daos.