This fixes a bizzare issue with deleting records from itertors.
It seems that in CF, if you don't name arguments they're passed into a method
and named 1,2, etc.  However, if you want to exploit that and manally name your attributes 1,2, etc
then the value passsed in will not make it into the method.  Doh!

In the end I just passed a record and added a hack with a named attribute.